Implement remainder of non-collection objects & attempt recursive
So this code includes the tests and return values of the rest of the possible basic
types of objects - the ones that aren't a collection themselves. I also took a stab
at wrapping the function in a special recursive function call that will return the
string 'null' for undefined results, since that is how JSON.stringify appears to
behave.

Next steps: begin working on collections and recursive calls.