Lay out pseudocode
I'm still not quite sure what the assigned task is for this one - both the details and the
high level leave me with a lot of confusion. On the high level, I'm maybe missing some
instructions or context. I see how this exercise integrates the DOM stuff that I've learned
in HTML & CSS lessons, but so far the objects returned by the function that I'm recreating
seem just way too densely packed for students to work on as an exercise. I think I might be
doing it wrong.

So all I did so far was write down some pseudocode for what I think this function is supposed
to do, in pretty vague terms. Part of it has some questions I think should be research priorities
that I think I should resolve pretty early on. I will start doing some reading on those.